apps/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Procesamiento documental ──────────────────────────────────────
openpyxl>=3.1.2
pandas>=2.2.0
xlrd>=2.0.1
PyMuPDF>=1.24.3
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.3.0
pytesseract>=0.3.10

# ── IA ────────────────────────────────────────────────────────────
openai>=1.30.0
anthropic>=0.26.0
requests>=2.31.0

# ── Generación de reportes ────────────────────────────────────────
python-pptx>=1.0.0
Jinja2>=3.1.4

# ── Seguridad ─────────────────────────────────────────────────────
cryptography>=43.0.0
keyring>=25.2.0

# ── Utilidades ────────────────────────────────────────────────────
python-multipart>=0.0.9
aiofiles>=23.2.1
